feat(types): add Moralis transaction and webhook types
- Introduced new TypeScript interfaces for Moralis transactions and webhook bodies, enhancing type safety and clarity in handling Moralis-related data.
- Updated the `copyToClipboard` function to improve reliability and user feedback, returning a boolean to indicate success or failure.
- Refactored the Moralis webhook processing to utilize the new types, ensuring better integration with the Moralis API.

This commit improves the application's structure for handling Moralis data and enhances the clipboard functionality for a better user experience.
